Renamed supercollider support functions
