+++ scsynth.js
